BosonScenario rewrite. We really have a design problem here - this is at least the third rewrite within a few weeks or so... But it shows that i'm doing something right: I am able to rewrite the entire BosonScenario class without the need to change a lot of the other classes. So at least the rest is good :-)
We have now a local XML document, instead of ScenarioPlayer and ScenarioUnit. We
load a file by simply copying the scenario node - same for saving. I am quite
happy with this class now. It seems that the game is still working and the
editor works for scenarios, too (finally). So I hope this was the last big
change.
We should be able to add anything we need now.

Last remark: maybe BosonScenario should be a friend of Unit/UnitBase. This way we don't
need to add that many public functions.