Those of you who have been on #boson recently know it already: for the last few weeks I have been working on OpenGL support. The bad news first: 3d is *not* yet supported. We also have a LOT of new bugs and missing features (e.g. select boxes are not yet integrated).
But the basic canvas draw routines are now rendered using OpenGL (in 2d) so we'll probably move to 3d in the near future.

Fell free to test this!

It is not yet working completely on all computers, thats why I am creating a branch instead of committing to HEAD as I planned initially.

You *need* 3d-accelleration in order to get the OpenGL working. I mean real hardware-accelleration, *not* software OpenGL! This won't work!

As a test try to use a) glxgears which isn't really useful as a test app (the gears should be moving very slow!) b) tuxkart c)tux-aqfh
Both of the games need hardware accellerated OpenGL support.

Have a lot of fun with this!