Just added a missile and a bullet (with teamcolor, you like that?). Maybe they are not correctly scaled compared to samsite/tank and others. Tell me if that is so.
Rivo: Both the bullet and the missile contain a small box that is invisible (because it's within another object), located near the rear of the objects and that has the name "particle_emitter". So if you ever feel like adding smoke trails and fire to the missile and/or a smoke trail to the bullet.... *g*

If this object is not needed, tell me and I'll remove it.

bmann and I agreed on creating a new dir called "objects" for these things. When I created that, I asked myself why we have a "particles" subdir. Would it not be a good idea to move those two pixmaps to "pixmaps"?

greetings,
	ben