- Boson::loadingStatus() is const - more meaningful (ahem - a *lot* of useful content now!) debug output for   delayed messages. usually they are simply advance messages - but when   something goes wrong we might need this - don't use KGame::setPolicy() in boson. we use PolicyClean *and* PolicyLocal   and setPolicy() would change policy of all of them. Instead we stream the   new value of gameStatus() and call KGamePropertyHandler::processMessage()   directly. - adding signalStartGameLoadWorkaround(). Well this too-long-name says   everything :)   the name emphasizes, that the code is not good. but now we can do this: - removed mLoading from TopWidget. Finally :-) now it resides in BosonStarting   only, and might even stay there.
starting games code should be designed much better now. the code still sucks
(mostly the same as it was before) but the design is *way* better. one day we
need to clean this code up, but this isn't so important anymore.

loading games should be fixed, now too - at least mostly. I can load a game
*and* move units around (you need to resize the window after loading, btw. this
is a minor bug, because of the changed design)