Created a Transall/C-130 or similar cargo-plane. May carry anything, but the idea was to create a realistic way to drop rivo's daisycutter, and to give andi a test-model for propeller-animation.
The plane is animated, it just opens the loading door and closes it again. The daisycutter rotates and opens the parachute. It does not fall.

Didn't watch the scale, might be completely off. A flash animation on how to drop the bomb is on http://www.usatoday.com/graphics/news/gra/daisycutter/frame.htm.

I am open for better names of the unit.

ben.