This model has a hundred frames (at least in 3dsmax :), which contain its construction animation. Its a 5 minute hack, so don't expect too much.
Should I do more of that, or are suddenly appearing objects from bottom-to-top better? How much memory will be used if every facility had a hundred frames?

Also, some buildings take longer to build than others. Can boson please deal with that, I just want to deliver a variable number of frames :)

Also, isn't there a redundancy between the isFacility-entry in the index-files and the mob/fix-prefix in the folder-names? Why is that so?

    ben