another version.
I never liked the fact that the facilities were on a sharply defined rectangle. Also, the facilities as they are right now, can be built right next to each other, which also looks stupid.

So this is an attempt to change it, which currently costs us 500 faces. But if we use, I can surely optimize it a little.

Right now its using the dirt texture, but try to imagine it with any other texture (i.e. cp concrt1.jpg rock1.jpg). I like that a lot better.