The samsite's missiles' paths are more straight now. Great would be if we could play a range of frames first, so the samsite can activate, open its doors and then fire.
rivo, leopard still doesn't use the bullet, did i mess it up in the config-files?

also, when i edit sam-sites index.unit to increase the speed value, the missiles aren't honoring that, they still have the same speed. why is that?

when boson loads, i get this line 9 times:
boson: WARNING: [virtual void BosonWeaponProperties::loadPlugin(KSimpleConfig*, bool)] Type is missile, but speed is 0, setting type to bullet

Also, the missiles trace is still a little off.

It would be great if fragments would bounce off the ground when they hit it, but thats exactly the same as the collision checking for hitting other units,
right?

greetings,
	ben