we cannot use QWidget::showFullScreen(), as we can't be sure that Qt has Xrandr support (mine doesn't). If we used it anyway we would end up in a window that is much larger than the current resolution.
but we can emulate showFullScreen() and give that emulation Xrandr support :-)

I tested this once only and it worked perfectly. this gives us a couple of nice
options - e.g. we can implement a usable "scroll when cursor is at the edge"
feature. and of course we can tell users to switch their resolution in the
options dialog when boson runs too slow :-)