Use a tree for generating the cell list. In a perfect world the original approach should be the perfect one, i.e. calculating the visible cells on the z=0 plane using mathematical formulas. Unfortunately this is all but easy and i believe we will never have acceptable code for this.
Therefore the tree solution is definitely easier to implement, is most probably
faster (definitely faster than current solution) and maybe even the best
solution existing at all.

So now we use a tree by default. The speed difference is significant, especially
when zoomed out greatly. On default zoom the difference is small, but on a
realistic zoom (zoomed out a bit, but not too far) it takes only about 20% of
the time of the original solution (0.6 ms compared to 3 ms. use a big map to
reproduce, 50x50 maps are not big enough).
Also the tree solution is even at a simple implementation O(logn) and therefore
still very fast with zoomed out completely on a 500x500 map (note that the FAR
plane will get in your way than). The perfect mathematical solution would be
O(1), but I believe I will never succeed at coding that (assuming it exists at
all), and the current mathematical solution is a LOT slower (you can notice
speed problems on high zoom levels with it. i could not anymore with the tree
solution).

Direct comparison can be made using different ground renderers: both, the
default and the fast one also provide a NoTree version.

I nominate this log for the longest-log-ever-for-small-commits prize!