this commit will cause more problems than it solves.
increasing the FAR plane from 100.0 to 1000.0
see comments in the header for why I think 1000 is a good number.

since the max camera z position dependet on the FAR plane I have changed that
too (i think 1000-50 as maximum z position is too much). it is now at 100, which
is twice as much as it was before.
this means you can zoom out much furhter, but it also means that you can get
into much higher speed problems than ever. I was able to reduce my FPS to < 4,
without a single unit being visible on a 500x500 map.