here it comes, the easy way to improve performance significantly, that I have searched for!!
after completing my LOD pathc it turned out that performance improved quite a
lot, but not as much as I had expected. the amount of time spent in model
rendering should be nearly linear to the number of meshes rendered. but even
with my patch, rendering a low number of units with many meshes (380 in this
case) took less time than a high number of units with few meshes (330 in 154
units in this case. this is 330 in _all_ units, not per unit).
I first thought that the mesh matrices were responsible, as they were the only
part of a mesh where still action happened, but it turned out they were not.

after a lot of profiling it turned out that there was a huge amount of time
spent in BosonModel::prepareRendering() (0.2 ms out of 0.3 ms for the complete
model rendering).
a closer look shows why: whenever a model is rendered, the vertex/texel/normal
arrays must be transferred to the GPU, no matter how many meshes are being
rendered. if you render 154 model/frame you need to ransfer 154 arrays to the
GPU (per frame).
the fix is easy: make sure that we re-use the array for as long as possible. the
easiest solution is to do that in the meshrenderer, the next step is to sort the
list of rendered items by model.

this patch gives me an improvement of 5-7 FPS on exp3 (zoomed out as far as
possible). doesn't sound much? it's an improvement of 50-70%, does that sound
better?
(exp3 without the patch: 10FPS, exp3 with the patch: 15-17FPS).

don't forget to do a make install to make this work!