the general consensus on boson-devel was that it's a good idea to provide a copy of KGame in our cvs, so here it comes. I copied the the whole kgame directory, without the dialogs/ subdir. we use only one class from there, which I copied in here, too. additionally I added kchatbase.* and kstdgameaction.* from libkdegames - this is necessary to remove the libkdegames dependency. I did not want to copy the while libkdegames.
boson links statically (as we do with all of our helper libraries) to this,
ignoring any libkdegames library that may be installed on your system. this
enables us to remove a couple from hacks that were made to keep boson working
with old libkdegames versions.

libkdegames (and therefore kdegames) is not required for boson anymore.