- pretty huge setLocalPlayer() redesign.   it has been a huge mess before (it has always been). now i think it is pretty   well designed.
  the localplayer is always NULL (except for the startup widgets) - it gets set
  once the game is started only.
  -> so the "set localplayer" task on startup is now just a special case of
     the "change localplayer" task in the editor.
- the startupwidgets now actually store the localplayer, where necessary, they
  don't touch boGame->localPlayer() anymore. boGame->localPlayer() will get
  removed soon, I hope.


this commit does significant changes to the startup procedure, which is very
sensitive to changes. I have not encountered new bugs yet (except for a problem
with the minimap - this might be related to this, or might be a rendering
problem), but I haven't tested all circumstances yet. please test as much as
possible - e.g. loading games ; multiple AI players ; editor mode ; ...
-> everything that influences the starting process