workaround for automake1.7 shared source files are now referenced by libraries only
this is inconvenient and ugly, but i see no other way. the proposed "solutions"
for the "[...] created both with libtool and without" error that I found don't
work, or cause other problems.
this solutions is pretty close to one of the proposed solutions (putting such
source files in dedicated libs), but it might cause other problems (e.g. with
later versions of automake), as they are still referenced by multiple libraries
(but not libs and executables anymore).