adding a memory manager for debugging
this code is supposed to trace every new or new[] as well as delete/delete[]
call and stores the amount of memory that was requested/freed. same happens for
malloc()/free() calls.

this is supposed to be a dependable way of finding memory consuming and maybe
leaking code. but the difference to the values reported by proc/pid/status are
so high that i am not sure whether this is actually bug free.

the memory manager adds a pretty big overhead (both, speed and memory) and is
therefore disabled by default. you need to configure with --enable-bomemory and
then do a "make clean; make" in order to use it