add a real event as action to BoCondition. currently this event is the only supported action for a condition, later we might add script functions as well.
the condition widget also supports this action event now.

using this we can already make winning conditions now, using the "PlayerLost"
event as action.
but since the GUI isn't very good yet you probably need to know what you're
doing (or at least be willing to ask a few details) to do that. fortunately the
GUI is the only bad thing right now, the data structures and the file format
seem to be sufficiently powerful.