This has been my next big project that I wanted to do. fortunately it was a lot easier to implement than I had expected. Unfortunately it will cause a lot of headache on us in the future ;-)
This is a network sync class, including a whole network sync protocol (slightly
inspired by popular network protocols). It does the task of checking for whether
the network is still in sync (i.e. all items are at the same position and state
on all clients) only.
It does _not_ fix a client that is out of sync.
(That will be another big project at a later time)

We are basically saving the states of all items and players into a
stream (not all states, only the most important ones, such as position, rotation
and health). This stream is just saved.
The ADMIN does this as well, but additionally he calculates an MD5 sum over this
stream and sends it out to the clients.
Once they receive this string, they compare their own data to this MD5 sum and
delete their stored data. The result of the comparison is sent back to the
ADMIN (an ACK).
The ADMIN waits for ACKs from all players and once he received all, he deletes
his data as well. If one of the clients tells him that it is out of sync, then
it compares that data of that client to his own data of that time, trying to
find the reason for the problem.
That reason (if found) is then printed as a chat message to the screen, so that
we can debug this.

It seems that we did a pretty good job with networking by now. I have just
completed a basic2 session without a problem.
You can also see that cheating causes sync trouble: use Debug->Unfog and start
some battles. You will very soon get out of sync.

Now - anyone brave enough to start a Storm session with network? *g*