removing MAXIMAL_ADVANCE_COUNT. All advance*() methods take the real advanceCallsCount as parameter now. most changes in this commit are simply advanceCount -> advanceCallsCount.
There is however, one feature that had to be removed because of this change:
previously the profiling dialog was able to group advance profilings by the
advance count number. using that we could find out in which advancecount most
time was wasted. since 2 different advance calls have always 2 different
advanceCallsCount values, this is not possible anymore.

The win of this however is that we have removed a whole level of complexity (the
distinction of advanceCount and advanceCallsCount is not necessary anymore)