- boson depends on plib >= 1.8.3 now. - adding an interface for plib classes. the interface provides signals and slots   and the API is pretty similar to Qt/boson API. - also adding a layout system that is slightly similar to QLayout (but much more   basic). the layout classes take BoPUI* widgets only (which in turn can contain   plib classes) - duplicating a couple of widgets - BoMatrixWidget and Bo*Camera* widgets now   provide both - the Qt/KDE implementation as well as a BoPUI (i.e. plib)   implementation.   this code duplication is temporary - the Qt/KDE classes will be removed once   everything has been ported
  note that while BoOrbiterCameraWidget has been ported, the BoOrbiterWidget has
  not yet been ported. it's kinda ironic, but porting the Orbiter widget is a
  lot of additional work (even though it is already OpenGL).
- converting the borender config widget (the control widgets on the left side)
  to BoPUI.
- temporarily disable compilation of bo3dsdebug, due to dependency issues.
  -> we need BosonGLWidget for bopui, however it is in libboson.la only

pui from plib is a very nice OpenGL widget toolkit, but note that it isn't as
mature as Qt. There are many widgets that simply don't exist and are non-trivial
to provide (such as a QTabWidget. I have made a very basic implementation only).
Also those that exist have some more or less minor problems that simply don't
exist in Qt, e.g. pressing the mouse button on a checkbox, moving the mouse away
and then releasing the button, sometimes has undefined (and even buggy) effects.