Supporting puMenuBar, the MenuBar of plib. It is a mostly complete implementation of something that is similar to KAction. Therefore the whole menubar thing is hidden completely to the actual menu designer. Everything is done (just like in KDE) with the *ui.rc files. The implementation also includes a BoPUIXMLBuilder class which is responsible for merging XML files, therefore we can use ui_standards.rc from kde without losing any features or any convenience. Our own XML merger class should also handle merging multiple files (and removing some of them later) better than KDE does, which comes from the fact that we don't need to provide several features that KDE has (e.g. the whole KParts idea is not used by boson, also startup time is a lot less important to us).
Note that only borender uses the new menubar and even there not all menu items
are included.

Limitations of the menubar:
- no KToggleAction equivalent in plib
- no nested menus supported by plib
- no toolbar class in plib
- every menu item must belong to an actual popupmenu, a menu item that is
  directly in the menubar is not supported by plib
- accels (such as 'q' for quit when a file menu is open) are not supported by
  plib. the actual accel can be easily implemented by us, however underlining
  the character is not so easy.

The toolbar limitations is not important to us, we can live without it. However
both, toggle actions and nested menus are very important to us. I will most
probably have to write our own plib based menubar rendering class :(

Additional problem:
an OpenGL menu requires that the menu is placed in the same context as the
other OpenGL elements, meaning we cannot use separat (actual) widgets, meaning
the menu takes space of the main widget, meaning the menu must be taken into
account when calculating the layout. However this makes things a lot more
difficult.