coyping puMenuBar.cxx over from plib and modifying it slightly to support nested menus. basically I give direct access to the underlying puPopupMenu pointer. by doing so, we can add more advanced items, not just const char* based text items.
note that this is a very trivial implementation that let's plib handly a lot of
things that it isn't supposed to handle. for example plib isn't supposed to open
a popupmenu for a nested menu.
however it works "acceptable" (even though it isn't nice), so I decided not to
spend time on this yet. but sooner or later i will have to improve it.

due to the design of our menubar code, not a single line outside of bopui/ had
to be modified