autoconf sucks. we cannot use advanced tests, because a) some systems don't support them the same way or b) some autoconf versions don't support them the same way
so we have to fall back to a minimalistic test. instead of actually linking
something useful against the plib libraries that we really need, we now just try
to link agains plibul, searching for ulInit.

I hope that won't matter on most systems, i.e. that we can assume that the
correct version of the library is at the correct location and that all relevant
libs are available.