adding a playfield test.
this test only cares about loading and saving right now. it first tries to load
a file, then tests whether it is valid, and finally tries to save it again and
compares the result to the previously loaded data.
the test fails for at least three of our maps, the loaded and the saved data
differs for some heightmaps.