terrain LOD.
previously the groundrenderer built a list of to-be-rendered cells whenever the
camera was moved - for this we used a quadtree.
That tree wasn't actually saved in memory, simple left/top/right/bottom integer
values were totally sufficient. Now we are actually constructing a tree.
However we don't save much additional data - only the integer coordinates and
number of cells at every node atm. This is because in editor mode all
interesting values (such as height of the cells) can change at any time and we
would have to re-generate these values.

However this simple tree in memory makes a few interesting features easier. We
can now easily traverse the tree at independant locations in the code - this is
necessary for usable level of detail.

So now whenever we find out that a node of the tree is completely in the view
frustum, we do not just add all cells, but check whether the distance of the
rectangle (to the NEAR plane) is high enough. If it is, we add the node itself,
instead of all its cells. I.e. instead of adding (x,y),(x,y+1),...,(x,y+n) we
add (x,y),(y+n) only. The rendering is done as if the cells in between would not
exist at all.

When a node is partially visible only, then we check for the LOD as well - if
the node is far away, then we add the whole node, otherwise we continue as
previously with the children.

The effects look pretty good to me (feel free to optimize
BoGroundRendererBase::doLOD() a bit - this is a trivial method and does the
important things).
However on maps with different heights there are often problems, because the
neighbour cells do not know that some cells may be missing. Fixing this is a bit
tricky unfortunately


btw: on experimental5, zoomed out completely with rotated camera, i get around
30 fps instead of 7 now