terrain LOD causes black edges under certain circumstances: e.g. in the case AAB AAC where A is a single rect with the size of 2*2 cells, and B and C are single cells. then the lower-left corner of B and top-left corner of C might have a height other then A at the same point. this is because only the top/bottom left/right corners of A are used, but not those in between.
this patch stores a separate heightmap for the groundrenderer that also contains
the correct values of the problematic corners of B/C.

memory usage is increased by about 1MB in worst case (500x500 map), rendering
speed should be unchanged.
however the generation of the celllist takes quite some additional time: first
the heightmap must be copied, then the value of the LOD rects must be calculated
and assigned to the new height map.

we are using a lazy strategy here, meaning a values of a cell are calculated
only if there is actually a cell that uses them. this adds some additional
overhead (some corners are calculated more than once) but it turns out that this
is a lot faster than other approaches.

heightmap copying takes about 5-10 ms in worst case (500x500 map, zoomed out and
rotated) currently.

i consider terrain LOD to be complete now, please let me know if problems come
up