as discussed on irc: adding an update counter to BosonEffect.
all effects have an update counter. in BosonCanvas::slotAdvance(), markUpdate()
is called which increases that counter, however the effect is not actually
updated (excpetion: effects that have not yet been started are updated
immediately).

in paintGL() we first make a list of visible effects and call update() on them
until their update counter reaches 0.
this way we avoid a lot of update() calls and hopefully reduces the amount of
time that is spent updating effects.

unfortunately this often increases the amount of time spent in paintGL(). an
improvement would be to update the visible effects whenever one of the following happens
- the camera changes (first update the list of visible effects, then update
  those)
- an advance call occurs (no update of the visible effects list required usually)