Something I wanted to do for a long time: adding a debug log.
The class BoDebugLog contains a method addEntry() that is called whenever a
debug message (debug/error/warning/fatal) is made. That message is then stored
in a log until the log reaches a certain size.
Currently the maximal entries count of debug messages is at 5,000 (this is very
low, but I believe that only the most recent messages are relevant anyway in
most cases), for errors and warnings 10,000 (pretty high, but if we ever reach
that limit with warnings or errors, then we have _serious_ trouble and start
fixing them).
Additionally there is a log that stores messages independantly of their type, to
keep the correct order (up to 10,000 messages).

The class can be configured to emit a signal when a message of a certain type
comes in. We use this currently for warnings and errors - they do now appear as
log messages on the screen. This may be very disturbing, but I don't see another
way to make people aware of the fact that there is a problem.

Additionally there is a debug dialog, that is able to read the whole log. You
can reach it through the debug menu.