Adding our own custom event loop. It takes care of delivering advance call events at the correct time to the game.
In contrast to the previous solution, this does not use a time, but rather
uses the actual current time and the desired call time. This is a lot more
dependable, as there is no lag from the timer timeout to the actual call.

Of course there is a lot of testing required for this, as it is completely new.
It seems to work very nice for me, but on Storm it still causes great problems
(I still believe they are caused by other parts of the game)