Fixed point data type. float/double are floating point data types, meaning they do not use a fixed number of bits for digits after the decimal point. The precision of these datatypes depend on the actual numbers. For example with floats you can perfectly store 4.0, 40.0, 0.1 and 4.1 However float can _not_ store 40.1 (you'll end up at 40.099998). Fixed point values on the other hand use a fixed number of bits for digits after the decimal point. This means that you often cannot represent a certain precision at all (for example 0.1 can't be represented with 17 bits after the point), but _if_ you can represent a value after the decimal point, then you can _always_ represent it, no matter what value is before the point.
The reason for introducing bofixed (the fixed point type) is network. Due to the
precision problem mentioned above (well mainly that one) floats are very
sensitive to precision loss. This isn't a problem for a network game if all
clients have the same precision loss at the same type (this is the case if all
use the same FPU). But if there is one client out there that does floating point
calculations slightly different (a different FPU - e.g. an old Athlon compared
to an Athlon XP has this problem) then the client will get out of sync very fast.

Fixed point arithmetics doesn't have that problem. This is because of the reason
described above, and because fixed point arithmetics can easily be implemented
using integers and without any FPU.

bofixed is a simple class implementing a fixed point data type. Most methods are
inline and it has only exactly one int member, so we should be able to use it
just like any primitive data type (int, float, ...).
bofixed uses 15 bits before the decimal point and 17 bits after the point. That
allows numbers from about -16384 to 16383.
Since we use bofixed mainly for item coordinates, we need values from 0 to 500
only (500 is the map size limit) so I believe this is completely sufficient.

The class is in the bomath.h header. This is supposed to be a very simple header
that doesn't depend on anything else (don't use BosonConfig or so here).

Nearly everything that network games depend on should be converted to bofixed
now - actually the conversion was mostly a matter of replacing float by bofixed.
Only sometimes explicit casting was necessary (mainly in foo?:bar1:bar2
statements). No semantic changes have been required at all.
bogameproperty.* provide two stream functions that are required when working
with KGameProperty<bofixed>. g++ cannot recognize that the property can be
casted to (and only to) bofixed, and bofixed to float. Therefore it refuses to
compile statements like
 boDebug() << property << endl;
where property is a KGameProperty<bofixed>.
By adding a simple implementation for operator<<() this can easily be fixed.
Just include bogameproperty.h instead of kgameproperty.h