restart the game timer when a client has left the game, when the local client is the server.
this has no effect if we had been server before (timer is just stopped and
started again), however if the client that left was the server, then this timer
restart enables the local client to continue the game (in single player mode).

this is one part of the do-not-crash-when-ADMIN-quits patch, however the other
patch is a patch against KGame and i want to commit it only once i ensured that
it won't break other KGame based games.