there is no point in making the SelectBoxData object static. instead we place it into the BosonBigDisplayBase object where it belongs (as it is used only there and as the GL context is there).
this fixes the (small) problem that the selection box display lists were not
deleted on destruction.

while I was on it I also removed the QMap that was used to store the display
lists. it doesn't make sense to use a map here, because all indices from list
base up to list base + list count - 1 are used anyway.