convert BoVector3/4 to templates.
I tried to use the Fixed version for all game logic variables and the Float one
for OpenGL-only variables. However sometimes conversions are needed and that's
the tricky part.
Especially when game logic methods use BoMatrix, which is float based. We use a
temporary Float vector that is later converted back to the Fixed version,
however it's not really nice.