KGameProperty IDs in Unit had been broken previously - they were dependant on the number of properties in UnitBase which can cause trouble when loading files from other versions.
now the IDs < 256 are reserved for KGame (as before), the IDs < 512 are reserved
for BosonItem (however it doesn't use any yet), the IDs < 1024 are reserved for
UnitBase and the rest is for Unit/UnitPlugins and derived classes.

all IDs have been adjusted and the file converted is included in this patch.