change class design to honor the fact that we use a OpenGL widget toolkit now.
this means that the RenderMain widget is mostly a dummy widget. it parses some
command line args, but everything else is done in the ModelPreview.

"Preview" is not a perfect name anymore for that thing, but since borender was
never meant (and still isn't) to be a perfectly clean application, I think we
can leave it as-is.

The Qt/KDE menubar has been removed. All other control widgets had been removed
before (i.e. replaced by libufo), this was the last thing to do.

Everything in borender is libufo based now, except for most dialogs. Some are
libufo based, but often still open a new window (instead of using the already
existing OpenGL context). But most are still normal dialogs. This may be fixed
"one day" but it is kinda low priority to me, as I think most important features
of libufo have been tested already.

I think we can start the conversion of boson to libufo.