converting the big display into a BosonUfoGLWidget. this is the first step of the libufo conversion. most renderText() code has been converted into libufo code (mostly labels). tooltips and chat messages are missing (these with their own classes for rendering).
atm both, ufo and "normal" text is rendered. you can disable this behavior by
setting UFO_LABELS_ONLY to 1, but currently the ufo text is hard to read
(black).

note that this patch assumes that ULabel (and therefore BoUfoLabel) supports
newlines. however that is not the case in 0.7.3 and not even in libufo cvs. i'll
try to get a patch on this in the next ufo release or otherwise we'll provide
the feature on our own.
so currently you will get some looooong lines where there should actually be
many short lines :)