ok, this one is both - evil and unexpected. we call addPropertyId() for every property in an item. this method adds the id along with its name to a static property map. because of that we cannot use the same Id in two different classes.
I am not sure whether this might have triggered bugs under certain
circumstances, but it definitely outputs errors (which nobody, including myself,
seems to have noticed before btw).

to fix this I have changed the ID areas of BosonShot, so that it
a) doesn't cover the areas of Unit anymore
b) derived classes don't cover areas of each other anymore.

this is very sad, because the previous solution was comparable easy while this
one reqires to adjust areas whenever a different derived classes is added. but
atm I don't see a good way around this.

this may cause recently saved games (including playfields) cause not to load
anymore. I did not write a converted for this, because it is a lot of work and
only very few maps should have that problem (the previous solution had never
been released and is relatively new, iirc).