adding bouic and boufodesigner.
the biggest problem that i've seen in converting from Qt/KDE widgets to OpenGL
widgets is the lack of a Qt designer equalivent. therefore only coders would be
able to create widgets.
my bodesigner (renamed to boufodesigner) project started out of curiosity
whether it'd be possible to create something similar (though much less powerful)
to to Qt designer / uic.

This commit does not include a great GUI designer, but rather the answer to that
question :-)

The boufodesigner is basically a simple widget that consists atm of 3 areas:
- the class list. this is a list of BoUfo classnames that can be placed on the
  screen
- the preview. this displays what your widget currently looks like
- the property editor. this allows you to edit properties of a widget inside
  the preview (e.g. text of a label, or type of the layout)

it is already mostly working as I'd like it to - you can place widgets into the
preview area, you can remove/move them and you can edit properties (especially
the layouts). the UI of this program sucks mostly - it is often required that
you know what you do in order to actually do it :)
the resulting widget can be saved into a .boui file which is just an xml file
describing the widget.

the bouic is the equivalent to Qt's user interface compiler uic. it takes a
.bouic file and creates a .h and a .cpp file out of it.
since we already know certain things about the widget (it's always a BoUfoWidget
subclass, boufo.h is always to be included, all members public, ...) we can
generate the .cpp/.h files in a pretty easy way without all of the difficult text
parsing that is usually required in compilers.

the entire thing works very well for me, as the welcomewidget.boui file proves.
it is a simple representation of the normal welcome widget as BoUfo widget.
there are several issues still outstanding (especially concerning layouts), but
these are general BoUfo issues. once they're fixed there, they're fixed in the
designer.

note that the .cpp/.h files are already generated and compiled, but not used
(and have not even been tested in an actual widget yet)