adding methods to save/load scripts to/from scripts.
we differ between saveScript() which just saves the normal .py file and
saveScriptData(), which saves the script as well as the current data of that
script (i.e. the values of the variables).

BosonScript does not yet provide a method to save the script itself, I hope
that'll be fixed soon (see the BOSONSCRIPT_LACKS_SAVE_SCRIPT define in
boeventlistener.cpp).

note that these methods are currently untested and unused, however that will be
fixed pretty soon