Adding a custom UI class for ULabels. This class is a copy of UBasicLabelUI with my line wrapping patch included.
I add this to our cvs because it seems that it won't get added to libufo
somewhen soon.

Note that this patch is not yet perfect - automatic line wrapping (e.g. required
for the chat widget) won't work because the layout system of libufo does not
take the actual size of the screen into account, i.e. a widget must tell it's
preferred size without knowing about the available size (of course it is
_preferred_ to a label not to wrap...)