Replacing our old KDE/Qt-widget based commandframe by a libufo based (i.e. OpenGL) commandframe.
There are some issues left, mainly about visual appearance. but the frame should
be usable.

The commandframe is now in BosonBigDisplayBase completely, not in
BosonWidgetBase (and derived classes), TopWidget, ... anymore. This is a great
improvement from a design pov, and actually it's been the reason for why I
wanted to start the "main" libufo port at the commandframe. The previous
solution has always been ugly and bad.

Since rewriting large parts was necessary anyway, this was a good opportunity to
slightly change some design apsects:
- There is no dedicated game/editor commandframe anymore. the commandframe has a
  setGameMode() method which hides the game/editor widgets (and
  disconnects/connects the relevant signals/slots).

  I am not absolutely convinced that this is a good idea, but it makes some
  things with libufo easier (and OpenGL in general), especially with out OpenGL
  widget design.
- The plugin structure (concerning the editor widgets and unit display widgets)
  has proven to be very useful and simple to use - improving this a bit more.
  E.g. all of these widgets now know whether they are visible in game or in
  editor mode and therefore we can use a simple loop to hide/show them according
  to the current game mode (previously we had to add custom code for every
  widget).