adding ufo menubar. toggle actions are not in sync with the kde menubar, but that one will get obsolete anyway
note that BosonBigDisplayBase now handles _way_ too much and therefore is a
total mess right now. this will get even worse before it'll get better again.
right now it is very hard to split the class up, as the final design is not yet
clear. this should get better once the other widgets (especially the startup
widgets) are ported, too.