Adding two macros that helped me quite a bit lately at profiling certain things:
* PROFILE_METHOD
This is a macro that creates a BosonProfiler object with a unique id for this
method. The method will appear in the profiling dialog with the same name that
appears when using k_funcinfo in debugging code.

* PROFILE_METHOD_2(name, desc)
This macro is essentially equal to the one above, but it also let's you specify
the name of the BosonProfiler object (so that you can use this macro more than
once in a method, or so that you can call name.stop() on it) and a description
that is appended to the method name in the profiling dialog.